rename block package to digest
