+++ crypto/digest/blocks.go
@@ -1,4 +1,4 @@
-package block
+package digest
 
 import (
 	"github.com/orbs-network/orbs-network-go/crypto/hash"
